services/identity-service/requirements-latest.txt
# Latest versions compatible with Python 3.13 (2024/2025)

# Web Framework
fastapi==0.116.1
uvicorn[standard]==0.32.1

# Data Validation
pydantic==2.11.7
pydantic-settings==2.7.1

# HTTP and Forms
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1

# Authentication & Security
PyJWT==2.10.1
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# Database (SQLAlchemy 2.0)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Caching
redis==5.2.1

# Email
aiosmtplib==3.0.2

# Messaging
aiokafka==0.11.0

# Monitoring & Observability
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# Service Discovery
python-consul==1.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
Faker==33.3.0

# Development
pre-commit==4.0.1